Fixed Range Bug
Everything works as its supposed to now. You have the basic sword set at
record 0 though, so basically the npcs will either drop Urethra or Basic
Sword insted of Urethra or Butthole

+++ final/Weapon.java
@@ -16,7 +16,7 @@ public class Weapon {
 	public int wid;
 	public String name;
 	public int damage, dur, range, rarity;	//dur: # of swings before it breaks; rate: # of shots per 5 seconds
-	public int rate;
+	public float rate;
 	public int ammo;
 //	private long lastUsed;
 	public Image model;
@@ -111,7 +111,7 @@ public class Weapon {
 	    	int minDam = raf.readInt();
 	    	int maxDam = raf.readInt();
 	    	this.damage = (int)((Math.random()*(maxDam-minDam))+minDam);
-	    	this.rate = raf.readInt();
+	    	this.rate = raf.readFloat();
 	    	this.range = raf.readInt();
 	    	this.dur = raf.readInt();
 	    	this.rarity = raf.readInt();
